+++ client/src/lib/json-to-iac/appid.js
@@ -1,4 +1,4 @@
-const { getObjectFromArray } = require("lazy-z");
+const { getObjectFromArray, splat, distinct } = require("lazy-z");
 const {
   rgIdRef,
   kebabName,
@@ -9,8 +9,49 @@ const {
   tfDone,
   jsonToTfPrint,
   getResourceOrData,
+  encryptionKeyRef,
+  cdktfRef,
+  getKmsInstanceData,
 } = require("./utils");
 const { varDotRegion } = require("../constants");
+
+/**
+ * create authorization policy to allow for appid to be encrypted
+ * @param {string} kmsName
+ * @param {Object} config configuration
+ * @returns {string} terraform code
+ */
+function ibmIamAuthorizationPolicyAppid(kmsName, config) {
+  let kmsInstance = getKmsInstanceData(kmsName, config);
+  return {
+    name: `appid to ${kmsName} kms policy`,
+    data: {
+      source_service_name: "appid",
+      roles: ["Reader"],
+      description: "Allow AppID instance to read from KMS instance",
+      target_service_name: kmsInstance.type,
+      target_resource_instance_id: cdktfRef(kmsInstance.guid),
+    },
+  };
+}
+
+/**
+ * format auth for appid to kms. appid authorization needs to be
+ * created prior to instance to allow for the appid service to be encrypted
+ * with keys from kms
+ * @param {string} kmsName
+ * @param {Object} config configuration
+ * @returns {string} terraform code
+ */
+function formatAppidToKmsAuth(kmsName, config) {
+  let auth = ibmIamAuthorizationPolicyAppid(kmsName, config);
+  return jsonToTfPrint(
+    "resource",
+    "ibm_iam_authorization_policy",
+    auth.name,
+    auth.data
+  );
+}
 
 /**
  * create terraform for appid key
@@ -67,11 +108,20 @@ function ibmResourceInstanceAppId(instance, config) {
   };
   // add needed values when new instance is created
   if (!instance.use_data) {
-    appIdValues.tags = config._options.tags;
     appIdValues.service = "appid";
     appIdValues.plan = "graduated-tier";
     appIdValues.location = varDotRegion;
   }
+  if (instance.kms) {
+    appIdValues.parameters = {
+      // paramters kms info need to be stringified json to work
+      kms_info: `{\\"id\\": \\"${cdktfRef(
+        getKmsInstanceData(instance.kms, config).guid
+      )}\\"}`,
+      tek_id: encryptionKeyRef(instance.kms, instance.encryption_key, "crn"),
+    };
+  }
+  if (!instance.use_data) appIdValues.tags = config._options.tags;
   return {
     name: instance.name,
     data: appIdValues,
@@ -131,6 +181,16 @@ function formatAppIdRedirectUrls(appid, urls, resourceName) {
  */
 function appidTf(config) {
   let tf = "";
+  let appidKmsInstances = distinct(splat(config.appid, "kms")).filter((kms) => {
+    // remove when undefined
+    if (kms) return kms;
+  });
+  let appidAuthPolicies = "";
+  appidKmsInstances.forEach((instance) => {
+    appidAuthPolicies += formatAppidToKmsAuth(instance, config);
+  });
+  if (appidAuthPolicies.length > 0)
+    tf += tfBlock(`Appid Authorization Policies`, appidAuthPolicies) + "\n";
   config.appid.forEach((instance) => {
     let str = formatAppId(instance, config);
     instance.keys.forEach((key) => {