+++ unit-tests/json-to-iac/appid.test.js
@@ -177,6 +177,83 @@ data "ibm_resource_instance" "test_appid" {
         "it should return correct data"
       );
     });
+    it("should format appid with encryption", () => {
+      let actualData = formatAppId(
+        {
+          name: "test-appid",
+          use_data: false,
+          resource_group: "slz-service-rg",
+          kms: "kms",
+          encryption_key: "key",
+        },
+        {
+          _options: {
+            prefix: "iac",
+            tags: ["hello", "world"],
+            region: "us-south",
+          },
+          resource_groups: [
+            {
+              use_prefix: false,
+              name: "slz-service-rg",
+              use_data: false,
+            },
+            {
+              use_prefix: false,
+              name: "slz-management-rg",
+              use_data: false,
+            },
+            {
+              use_prefix: false,
+              name: "slz-workload-rg",
+              use_data: false,
+            },
+          ],
+          key_management: [
+            {
+              name: "kms",
+              service: "kms",
+              resource_group: "slz-service-rg",
+              authorize_vpc_reader_role: true,
+              use_data: false,
+              use_hs_crypto: false,
+              keys: [
+                {
+                  name: "key",
+                  root_key: true,
+                  key_ring: "test",
+                  force_delete: true,
+                  rotation: 12,
+                  dual_auth_delete: true,
+                },
+              ],
+            },
+          ],
+        }
+      );
+      let expectedData = `
+resource "ibm_resource_instance" "test_appid" {
+  name              = "\${var.prefix}-test-appid"
+  resource_group_id = ibm_resource_group.slz_service_rg.id
+  service           = "appid"
+  plan              = "graduated-tier"
+  location          = var.region
+  parameters = {
+    kms_info = "{\\"id\\": \\"\${ibm_resource_instance.kms.guid}\\"}"
+    tek_id   = ibm_kms_key.kms_key_key.crn
+  }
+  tags = [
+    "hello",
+    "world"
+  ]
+}
+`;
+      assert.deepEqual(
+        actualData,
+        expectedData,
+        "it should return correct data"
+      );
+    });
   });
   describe("formatAppIdRedirectUrls", () => {
     it("should format appid urls", () => {
@@ -283,6 +360,134 @@ resource "ibm_resource_instance" "test_appid" {
   service           = "appid"
   plan              = "graduated-tier"
   location          = var.region
+  tags = [
+    "hello",
+    "world"
+  ]
+}
+
+resource "ibm_resource_key" "test_appid_key_test_key" {
+  name                 = "\${var.prefix}-test-appid-test-key"
+  resource_instance_id = ibm_resource_instance.test_appid.id
+  role                 = "Writer"
+  tags = [
+    "hello",
+    "world"
+  ]
+}
+
+resource "ibm_resource_key" "test_appid_key_test_key_2" {
+  name                 = "\${var.prefix}-test-appid-test-key-2"
+  resource_instance_id = ibm_resource_instance.test_appid.id
+  role                 = "Writer"
+  tags = [
+    "hello",
+    "world"
+  ]
+}
+
+##############################################################################
+`;
+      assert.deepEqual(
+        actualData,
+        expectedData,
+        "it should return correct data"
+      );
+    });
+    it("should create appid terraform with encrypted appid instances", () => {
+      let actualData = appidTf({
+        _options: {
+          prefix: "iac",
+          tags: ["hello", "world"],
+          region: "us-south",
+        },
+        appid: [
+          {
+            name: "test-appid",
+            use_data: false,
+            resource_group: "slz-service-rg",
+            keys: [
+              {
+                appid: "test-appid",
+                name: "test-key",
+              },
+              {
+                appid: "test-appid",
+                name: "test-key-2",
+              },
+            ],
+            kms: "kms",
+            encryption_key: "key",
+          },
+        ],
+        resource_groups: [
+          {
+            use_prefix: false,
+            name: "slz-service-rg",
+            use_data: false,
+          },
+          {
+            use_prefix: false,
+            name: "slz-management-rg",
+            use_data: false,
+          },
+          {
+            use_prefix: false,
+            name: "slz-workload-rg",
+            use_data: false,
+          },
+        ],
+        key_management: [
+          {
+            name: "kms",
+            service: "kms",
+            resource_group: "slz-service-rg",
+            authorize_vpc_reader_role: true,
+            use_data: false,
+            use_hs_crypto: false,
+            keys: [
+              {
+                name: "key",
+                root_key: true,
+                key_ring: "test",
+                force_delete: true,
+                rotation: 12,
+                dual_auth_delete: true,
+              },
+            ],
+          },
+        ],
+      });
+      let expectedData = `##############################################################################
+# Appid Authorization Policies
+##############################################################################
+
+resource "ibm_iam_authorization_policy" "appid_to_kms_kms_policy" {
+  source_service_name         = "appid"
+  description                 = "Allow AppID instance to read from KMS instance"
+  target_service_name         = "kms"
+  target_resource_instance_id = ibm_resource_instance.kms.guid
+  roles = [
+    "Reader"
+  ]
+}
+
+##############################################################################
+
+##############################################################################
+# Test Appid Resources
+##############################################################################
+
+resource "ibm_resource_instance" "test_appid" {
+  name              = "\${var.prefix}-test-appid"
+  resource_group_id = ibm_resource_group.slz_service_rg.id
+  service           = "appid"
+  plan              = "graduated-tier"
+  location          = var.region
+  parameters = {
+    kms_info = "{\\"id\\": \\"\${ibm_resource_instance.kms.guid}\\"}"
+    tek_id   = ibm_kms_key.kms_key_key.crn
+  }
   tags = [
     "hello",
     "world"